Spec Generation WIP
This is a WIP to generate table spec files.

It pulls out some stuff from `table.go` to `column.go`

To generate specs, we need to expand the options bitmask into the set of booleans. Instead of doing that via the bitmask (as in #77), this explores storing these as booleans and generating the bitmask as needed.