Use ranges on timing related tests
I've noticed several failures on these tests. Generally, I think it's
because GitHub Actions is on shared CPUs and prone to randomly pausing.
This, in turn, makes the timing tests very squirrely

This is an attempt at converting from expected counts, to range checks
I don't love it, and it's not going to catch all failures, but I'm
not sure there's a better approach